requirements.txt
torch>=2.0.0
numpy>=1.24.0
pygame>=2.5.0
pyyaml>=6.0
tensorboard>=2.14.0
pillow>=10.0.0
